+++ lib/packages/wikibase/data-model/tests/unit/Term/AliasGroupListTest.php
@@ -249,7 +249,7 @@ class AliasGroupListTest extends \PHPUnit_Framework_TestCase {
 	public function testGivenAliasGroupArgs_setGroupTextsSetsAliasGroup() {
 		$list = new AliasGroupList();
 
-		$list->setGroupTexts( 'en', array( 'foo', 'bar' ) );
+		$list->setAliasesForLanguage( 'en', array( 'foo', 'bar' ) );
 
 		$this->assertEquals(
 			new AliasGroup( 'en', array( 'foo', 'bar' ) ),
@@ -261,14 +261,14 @@ class AliasGroupListTest extends \PHPUnit_Framework_TestCase {
 		$list = new AliasGroupList();
 
 		$this->setExpectedException( 'InvalidArgumentException' );
-		$list->setGroupTexts( null, array( 'foo', 'bar' ) );
+		$list->setAliasesForLanguage( null, array( 'foo', 'bar' ) );
 	}
 
 	public function testGivenInvalidAliases_setGroupTextsThrowsException() {
 		$list = new AliasGroupList();
 
 		$this->setExpectedException( 'InvalidArgumentException' );
-		$list->setGroupTexts( 'en', array( 'foo', null ) );
+		$list->setAliasesForLanguage( 'en', array( 'foo', null ) );
 	}
 
 }